♻️ Move SQLite code in its own directory
